@@ -9,8 +9,8 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
+    "generate": "amplify codegen ",
     "build": "ng build --prod",
-    "codegen": "",
     "test": "jest",
     "test:watch": "jest --watch",
     "lint": "ng lint",
@@ -18,38 +18,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.0",
-    "@angular/cdk": "^8.0.0",
-    "@angular/common": "^8.0.0",
-    "@angular/compiler": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/forms": "^8.0.0",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@angular/router": "^8.0.0",
+    "@angular/animations": "^8.0.2",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "^8.0.2",
+    "@angular/compiler": "^8.0.2",
+    "@angular/core": "^8.0.2",
+    "@angular/forms": "^8.0.2",
+    "@angular/platform-browser": "^8.0.2",
+    "@angular/platform-browser-dynamic": "^8.0.2",
+    "@angular/router": "^8.0.2",
     "alife-rating-star": "0.0.1",
-    "aws-amplify": "^1.1.28",
-    "aws-amplify-angular": "^3.0.3",
+    "aws-amplify": "^1.1.29",
+    "aws-amplify-angular": "^3.0.4",
     "bulma": "^0.7.5",
     "bulma-slider": "^2.0.0",
     "bulma-tagsinput": "^2.0.0",
-    "core-js": "^2.6.9",
+    "core-js": "^3.1.4",
     "font-awesome": "^4.7.0",
     "moment": "^2.24.0",
     "rxjs": "^6.5.2",
-    "typescript": "^3.4.5",
-    "webpack": "^4.32.2",
     "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.800.1",
-    "@angular/cli": "^8.0.1",
-    "@angular/compiler-cli": "^8.0.0",
-    "@angular/language-service": "^8.0.0",
-    "@types/jest": "^24.0.13",
+    "@angular-devkit/build-angular": "^0.800.3",
+    "@angular/cli": "^8.0.3",
+    "@angular/compiler-cli": "^8.0.2",
+    "@angular/language-service": "^8.0.2",
+    "@types/jest": "^24.0.15",
     "@types/mocha": "^5.2.7",
-    "@types/node": "^12.0.4",
-    "codelyzer": "^5.0.1",
+    "@types/node": "^12.0.10",
+    "codelyzer": "^5.1.0",
     "flatted": "^2.0.0",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -57,7 +55,9 @@
     "jest-preset-angular": "^7.1.1",
     "jsdom": "^15.1.1",
     "protractor": "^5.4.2",
-    "ts-node": "~8.2.0",
-    "tslint": "~5.17.0"
+    "ts-node": "~8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "3.4.5",
+    "webpack": "^4.35.0"
   }
 }
