@@ -1,8 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
-import {RecipeService} from '../../Services';
 import * as moment from 'moment';
-const weekDays = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday'];
+import {APIService, BatchGetRecipesQuery} from '../../../API.service';
 @Component({
   selector: 'app-weeklist',
   templateUrl: './weekList.html',
@@ -13,9 +12,9 @@ export class WeeklistComponent implements OnInit {
   week: number;
   year: number;
   weekDates;
-  weekRecipes: Promise<any[]>;
-
-  constructor(private recipeService: RecipeService, private route: ActivatedRoute,
+  private weekRecipes: Promise<any>;
+  constructor(private route: ActivatedRoute,
+              private apiService: APIService,
               public router: Router) {
     this.route.params.subscribe(res => {
       this.week = res.week;
@@ -25,32 +24,16 @@ export class WeeklistComponent implements OnInit {
 
   ngOnInit() {
     this.weekDates = this.setDatesInWeek();
-    this.weekRecipes = this.getFakedRandomWeek();
-    // this.getRandomWeek();
+    this.weekRecipes = this.getRandomWeek();
   }
-  getFakedRandomWeek() {
-    this.submitted = false;
-    return this.recipeService.searchRecipes('pizza').then(res  => {
-       return res.map((recipe, index) => {
-         recipe.recipe.servings = 4;
-         recipe.recipe.weekDay = weekDays[index];
-         return recipe.recipe;
-        }).slice(0, 7);
-    });
-  }
-
-  getRandomWeek(week, year) {
-    // this.AP.getWeekList(week, year).subscribe((ret) => {
-    //   this.submitted = ret.final;
-    //   this.weekRecipes = this.recipeService.getRecipes(ret.recipes);
-    // });
+  getRandomWeek() {
+    return this.apiService.GetMenu('' + this.week + this.year);
   }
 
   navigate(weeks) {
     const date = moment().week(this.week).year(this.year);
     date.add(weeks, 'week');
     this.weekDates = this.setDatesInWeek();
-    // this.getRandomWeek();
     this.router.navigate([`../menu/${date.get('year')}/${date.get('week')}`]);
   }
 
@@ -65,6 +48,4 @@ export class WeeklistComponent implements OnInit {
     return dates;
   }
 
-  verifyWeekList() {
-  }
 }
